issue-104: defined an interface for the fa_role functionality.
